+++ test/DataStructures/ClosedMap_uut.h
@@ -22,7 +22,7 @@ TEST_CASE("ClosedMap test"){
     SECTION("Searching for object"){
         REQUIRE(closedMap.getSize() == 16);
         for(std::size_t i = 0; i < 16; i++){
-            REQUIRE(closedMap.at(std::to_string(i)).value().get().m_Id == i);
+            REQUIRE(closedMap.at(std::to_string(i)).value().get().m_Id == static_cast<int>(i));
         }
         REQUIRE(closedMap.at("-1").has_value() == false);
         REQUIRE(closedMap.at("100").has_value() == false);
@@ -38,7 +38,7 @@ TEST_CASE("ClosedMap test"){
             REQUIRE(closedMap.contains(std::to_string(i)) == false);
         }
         for(std::size_t i = 4; i < 16   ; i++){
-            REQUIRE(closedMap.at(std::to_string(i)).value().get().m_Id == i);
+            REQUIRE(closedMap.at(std::to_string(i)).value().get().m_Id == static_cast<int>(i));
         }
     }
